new Features & Bugfixes
- Automatic UTC-Conversion Attribute when DateTime values are stored as UTC in a database
- Generic Adapters for Configuration Exchange between databases
- new Features in IDbSet (new record, read index-value of an entity and dynamic querying
- PluginFactory support for Generic Plugin-Types
- Authentication Features now also work without explicit Authentication Scheme. If not specified, the requirements are checked for all authentication types
- Feature-Driven security
- Bugfix: Task.Fulfill can be called subsequently without causing an error
- ChangeType callback in DefaultCallbacks of scripting
- MemberExtensions now support GetValueType for public properties and fields
- All Contexts for Web-Seucirty supported by the default-views
- new Security-Features for ForeignKey selection
- TenantUsers can now be disabled without the need to delete them
- New Validation Attributes
  - ConditionalRequired for values that are only required when a specific condition is met
  - ForceTrue for switches or check-boxes that need to be checked
- Tenant-Binding in Default-Views can be switched for selection using multi-select control or with a sub-grid
- ShowTable component now support additional views
- ExternalProgramInvoker can now be used async

+++ ITVComponents.WebCoreToolkit.Net.TelerikUI.TSV/Areas/Security/Controllers/NavigationController.cs
@@ -1,20 +1,23 @@
-﻿using System.Linq;
+﻿using System;
+using System.Linq;
 using System.Threading.Tasks;
 using ITVComponents.DataAccess.Extensions;
 using ITVComponents.WebCoreToolkit.AspExtensions;
 using ITVComponents.WebCoreToolkit.EntityFramework.TenantSecurityShared;
 using ITVComponents.WebCoreToolkit.EntityFramework.TenantSecurityShared.Models.Base;
 using ITVComponents.WebCoreToolkit.MvcExtensions;
+using ITVComponents.WebCoreToolkit.Net.TelerikUi.TenantSecurityViews.Options;
 using ITVComponents.WebCoreToolkit.Net.TelerikUi.TenantSecurityViews.ViewModel;
 using Kendo.Mvc.Extensions;
 using Kendo.Mvc.UI;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Options;
 
 namespace ITVComponents.WebCoreToolkit.Net.TelerikUi.TenantSecurityViews.Areas.Security.Controllers
 {
-    [Authorize(Policy="HasPermission(Navigation.View,Navigation.Write)"), Area("Security"), ConstructedGenericControllerConvention]
+    [Authorize(Policy= "HasPermission(Navigation.View,Navigation.Write),HasFeature(ITVAdminViews)"), Area("Security"), ConstructedGenericControllerConvention]
     public class NavigationController<TUserId, TUser, TRole, TPermission, TUserRole, TRolePermission, TTenantUser, TNavigationMenu, TTenantNavigation, TQuery, TQueryParameter, TTenantQuery, TWidget, TWidgetParam, TUserWidget, TUserProperty, TContext> : Controller
         where TRole : Role<TUserId, TUser, TRole, TPermission, TUserRole, TRolePermission, TTenantUser>
         where TPermission : Permission<TUserId, TUser, TRole, TPermission, TUserRole, TRolePermission, TTenantUser>
@@ -34,10 +37,13 @@ namespace ITVComponents.WebCoreToolkit.Net.TelerikUi.TenantSecurityViews.Areas.S
         where TContext: DbContext,ISecurityContext<TUserId,TUser,TRole,TPermission,TUserRole,TRolePermission,TTenantUser,TNavigationMenu,TTenantNavigation,TQuery,TQueryParameter,TTenantQuery,TWidget,TWidgetParam,TUserWidget,TUserProperty>
     {
         private readonly TContext db;
+        private readonly IOptions<SecurityViewsOptions> options;
 
-        public NavigationController(TContext db)
+        public NavigationController(TContext db, IOptions<SecurityViewsOptions> options)
         {
             this.db = db;
+            this.options = options;
+            db.ShowAllTenants = true;
         }
 
         public IActionResult Index()
@@ -53,10 +59,17 @@ namespace ITVComponents.WebCoreToolkit.Net.TelerikUi.TenantSecurityViews.Areas.S
             return PartialView();
         }
 
+
+        public IActionResult NavDetailTabs(int parentId)
+        {
+            ViewData["parentId"] = parentId;
+            //ViewData["Permissions"] = db.ReadForeignKey("Permissions").Cast<ForeignKeyData<int>>().ToList();
+            return PartialView();
+        }
+
         [HttpPost]
         public IActionResult Read([DataSourceRequest] DataSourceRequest request, [FromQuery]int? parentId)
         {
-            db.ShowAllTenants = true;
             return Json(db.Navigation.Where(n => n.ParentId == parentId).ToDataSourceResult(request, n => n.ToViewModel<TNavigationMenu, NavigationMenuViewModel>(ApplyTenants)));
         }
 
@@ -71,6 +84,7 @@ namespace ITVComponents.WebCoreToolkit.Net.TelerikUi.TenantSecurityViews.Areas.S
                 await this.TryUpdateModelAsync<NavigationMenuViewModel,TNavigationMenu>(model);
                 model.ParentId = parentId;
                 model.SortOrder = order + 1;
+                model.RefTag = Guid.NewGuid().ToString("D");
                 db.Navigation.Add(model);
                 SetTenants(viewModel.Tenants, model);
                 await db.SaveChangesAsync();
@@ -83,7 +97,6 @@ namespace ITVComponents.WebCoreToolkit.Net.TelerikUi.TenantSecurityViews.Areas.S
         [Authorize(Policy="HasPermission(Navigation.Write)")]
         public async Task<IActionResult> Destroy([DataSourceRequest] DataSourceRequest request, NavigationMenuViewModel viewModel)
         {
-            db.ShowAllTenants = true;
             var model = db.Navigation.First(n => n.NavigationMenuId == viewModel.NavigationMenuId);
             if (ModelState.IsValid)
             {
@@ -99,7 +112,6 @@ namespace ITVComponents.WebCoreToolkit.Net.TelerikUi.TenantSecurityViews.Areas.S
         [Authorize(Policy="HasPermission(Navigation.Write)")]
         public async Task<IActionResult> Update([DataSourceRequest] DataSourceRequest request, NavigationMenuViewModel viewModel)
         {
-            db.ShowAllTenants = true;
             var model = db.Navigation.First(n => n.NavigationMenuId == viewModel.NavigationMenuId);
             if (ModelState.IsValid)
             {
@@ -116,25 +128,36 @@ namespace ITVComponents.WebCoreToolkit.Net.TelerikUi.TenantSecurityViews.Areas.S
 
         private void ApplyTenants(TNavigationMenu original, NavigationMenuViewModel vm)
         {
-            vm.Tenants = (from t in original.Tenants select t.TenantId).ToArray();
+            if (options.Value.TenantLinkMode == LinkMode.MultiSelect)
+            {
+                vm.Tenants = (from t in original.Tenants select t.TenantId).ToArray();
+            }
+            else
+            {
+                vm.Tenants = Array.Empty<int>();
+            }
         }
 
         private void SetTenants(int[] tenants, TNavigationMenu menu)
         {
-            var tmp = (from t in menu.Tenants
-                join nt in tenants on t.TenantId equals nt
-                    into j
-                from o in j.DefaultIfEmpty()
-                where o == 0
-                select t).ToArray();
-            var tmp2 = (from t in tenants
-                join ot in menu.Tenants on t equals ot.TenantId
-                    into j
-                from n in j.DefaultIfEmpty()
-                where n == null
-                select t).ToArray();
-            db.TenantNavigation.RemoveRange(tmp);
-            db.TenantNavigation.AddRange(from t in tmp2 select new TTenantNavigation {TenantId = t, NavigationMenu = menu});
+            if (options.Value.TenantLinkMode == LinkMode.MultiSelect)
+            {
+                var tmp = (from t in menu.Tenants
+                    join nt in tenants on t.TenantId equals nt
+                        into j
+                    from o in j.DefaultIfEmpty()
+                    where o == 0
+                    select t).ToArray();
+                var tmp2 = (from t in tenants
+                    join ot in menu.Tenants on t equals ot.TenantId
+                        into j
+                    from n in j.DefaultIfEmpty()
+                    where n == null
+                    select t).ToArray();
+                db.TenantNavigation.RemoveRange(tmp);
+                db.TenantNavigation.AddRange(from t in tmp2
+                    select new TTenantNavigation { TenantId = t, NavigationMenu = menu });
+            }
         }
     }
 }
